Moved Control Flag to its own file.
